float  to  double

+++ RLS.cpp
@@ -9,12 +9,12 @@ RLS::RLS(int _M, RLS_TYPE_e _type, float _rho, int _p)
     H_.setZero();
     Y_.setZero();
     rho_p_ = std::pow(rho_, p_);
-    P_ = Eigen::MatrixXf::Identity(M_, M_) * 1024;
+    P_ = Eigen::MatrixXd::Identity(M_, M_) * 1024;
 }
 
 RLS::~RLS() {}
 
-Eigen::VectorXf RLS::initRegression(const Eigen::MatrixXf& _H, const Eigen::MatrixXf& _Y)
+Eigen::VectorXd RLS::initRegression(const Eigen::MatrixXd& _H, const Eigen::MatrixXd& _Y)
 {
     H_.bottomRows(_H.rows()) = _H;
     Y_.bottomRows(_Y.rows()) = _Y;
@@ -36,14 +36,14 @@ Eigen::VectorXf RLS::initRegression(const Eigen::MatrixXf& _H, const Eigen::Matr
     return theta_;
 }
 
-Eigen::VectorXf RLS::recursiveRegression(const Eigen::RowVectorXf& _new_h, const Eigen::RowVectorXf& _new_y)
+Eigen::VectorXd RLS::recursiveRegression(const Eigen::RowVectorXd& _new_h, const Eigen::RowVectorXd& _new_y)
 {
     if (!is_init_)
     {
         return initRegression(_new_h, _new_y);
     }
 
-    Eigen::MatrixXf Q = (P_ - (P_ * _new_h.transpose() * _new_h * P_) / (rho_ + _new_h * P_ * _new_h.transpose())) / rho_;
+    Eigen::MatrixXd Q = (P_ - (P_ * _new_h.transpose() * _new_h * P_) / (rho_ + _new_h * P_ * _new_h.transpose())) / rho_;
     P_ = Q + (rho_p_ * Q * H_.row(0).transpose() * H_.row(0) * Q) / (1.0f - rho_p_ * H_.row(0) * Q * H_.row(0).transpose());
     theta_ = theta_ + P_ * _new_h.transpose() * (_new_y - _new_h * theta_)
         - P_ * rho_p_ * H_.row(0).transpose() * (Y_.row(0) - H_.row(0) * theta_);
